:lipstick: - style: move the re-submit button to the secondary navigation
